smart-assistance1-main/requirements.txt
streamlit
PyMuPDF
nltk
sentence-transformers
faiss-cpu
transformers
